make linkme optional
